Add base find class to reduce code duplication in many api classes that utilize find.
