+++ lib/tankard/api/adjunct.rb
@@ -1,5 +1,5 @@
 require 'hashie'
-require 'tankard/api/utils/find'
+require 'tankard/api/base/find'
 
 module Tankard
   module Api
@@ -7,18 +7,8 @@ module Tankard
     #
     # @see http://www.brewerydb.com/developers/docs-endpoint/adjunct_index
     # @author Matthew Shafer
-    class Adjunct
-      include Tankard::Api::Utils::Find
-
-      def initialize(request, options = {})
-        @http_client = request
-        @http_request_parameters = Hashie::Mash.new(options)
-      end
-
+    class Adjunct < Tankard::Api::Base::Find
     private
-
-      attr_reader :http_client
-      attr_reader :http_request_parameters
 
       def route
         'adjunct'
