+++ lib/tankard/api/category.rb
@@ -1,4 +1,4 @@
-require 'tankard/api/utils/find'
+require 'tankard/api/base/find'
 
 module Tankard
   module Api
@@ -6,19 +6,13 @@ module Tankard
     #
     # @see http://www.brewerydb.com/developers/docs-endpoint/category_index
     # @author Matthew Shafer
-    class Category
-      include Tankard::Api::Utils::Find
-
-      # Initializes a new object
+    class Category < Tankard::Api::Base::Find
+      # @!method initialize(request, options = {})
+      #   Initializes a new object
       #
-      # @param request [Tankard::Request]
-      # @param options [Hash]
-      # @return [Tankard::Api::Category]
-      def initialize(request, options = {})
-        @http_client = request
-        @http_request_parameters = options
-        @route = 'category'
-      end
+      #   @param request [Tankard::Request]
+      #   @param options [Hash]
+      #   @return [Tankard::Api::Category]
 
       # @!method find(id_or_array, options={})
       #   Find a single or multiple categories by their id
@@ -31,7 +25,9 @@ module Tankard
 
     private
 
-      attr_reader :http_client, :http_request_parameters, :route
+      def route
+        'category'
+      end
     end
   end
 end
