+++ lib/tankard/api/style.rb
@@ -1,4 +1,4 @@
-require 'tankard/api/utils/find'
+require 'tankard/api/base/find'
 
 module Tankard
   module Api
@@ -6,19 +6,13 @@ module Tankard
     #
     # @see http://www.brewerydb.com/developers/docs-endpoint/style_index
     # @author Matthew Shafer
-    class Style
-      include Tankard::Api::Utils::Find
-
-      # Initializes a new object
+    class Style < Tankard::Api::Base::Find
+      # @!method initialize(request, options = {})
+      #   Initializes a new object
       #
-      # @param request [Tankard::Request]
-      # @param options [Hash]
-      # @return [Tankard::Api::Style]
-      def initialize(request, options = {})
-        @http_client = request
-        @http_request_parameters = options
-        @route = 'style'
-      end
+      #   @param request [Tankard::Request]
+      #   @param options [Hash]
+      #   @return [Tankard::Api::Style]
 
       # @!method find(id_or_array, options={})
       #   Find a single or multiple styles by their id
@@ -28,12 +22,12 @@ module Tankard
       #   @return [Hash, Array] if a string with a style id is passed to find then the hash of the style is returned.
       #     if an array is passed to find an array containing hashes with each style is returned.
       #     if a style is not found nothing for that style is returned.
-
+      
     private
 
-      attr_reader :http_client
-      attr_reader :http_request_parameters
-      attr_reader :route
+      def route
+        'style'
+      end
     end
   end
 end
