+++ lib/tankard/api/yeast.rb
@@ -1,4 +1,4 @@
-require 'tankard/api/utils/find'
+require 'tankard/api/base/find'
 
 module Tankard
   module Api
@@ -6,19 +6,13 @@ module Tankard
     #
     # @see http://www.brewerydb.com/developers/docs-endpoint/yeast_index
     # @author Matthew Shafer
-    class Yeast
-      include Tankard::Api::Utils::Find
-
-      # Initializes a new object
+    class Yeast < Tankard::Api::Base::Find
+      # @!method initialize(request, options = {})
+      #   Initializes a new object
       #
-      # @param request [Tankard::Request]
-      # @param options [Hash]
-      # @return [Tankard::Api::Yeast]
-      def initialize(request, options = {})
-        @http_client = request
-        @http_request_parameters = options
-        @route = 'yeast'
-      end
+      #   @param request [Tankard::Request]
+      #   @param options [Hash]
+      #   @return [Tankard::Api::Yeast]
 
       # @!method find(id_or_array, options={})
       #   Find a single or multiple yeasts by their id
@@ -31,9 +25,9 @@ module Tankard
 
     private
 
-      attr_reader :http_client
-      attr_reader :http_request_parameters
-      attr_reader :route
+      def route
+        'yeast'
+      end
     end
   end
 end
